requirements.txt
flask
python-dotenv
twilio
openai
requests
Pillow




